requirements.txt
flask
openai
gunicorn
pdfplumber
pdf2image
pytesseract
Pillow
pillow-heif
anthropic
